requirements.txt
pyspark==3.1.2
netCDF4==1.5.7
gcsfs==2021.11.1
google-cloud-storage==1.43.0
pandas==1.3.4
h3==3.7.3
